agents/analyzer_agent/requirements.txt
uagents>=0.9.2
openai>=1.12.0
python-dotenv>=1.0.0
